requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
